backend/pyproject.toml
[tool.poetry]
name = "switchkeys"
version = "0.1.0"
description = ""
authors = ["Mahmoud Emad <mahmmoud.hassanein@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.1"
python-decouple = "^3.8"
django-cors-headers = "^4.3.1"
pillow = "^10.2.0"
drf-yasg = "^1.21.7"
django-debug-toolbar = "^4.3.0"
black = "^24.3.0"
flake8 = "^7.0.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
